:truck: moved /collections to the CollectionList
